+++ tibiapy/__init__.py
@@ -17,7 +17,7 @@ from tibiapy.world import *
 from tibiapy.bazaar import *
 from tibiapy.client import *
 
-__version__ = '3.3.0'
+__version__ = '3.4.0'
 
 from logging import NullHandler