+++ CHANGELOG.rst
@@ -6,6 +6,12 @@ Changelog
     Due to this library relying on external content, older versions are not guaranteed to work.
     Try to always use the latest version.
 
+3.4.0 (2020-09-19)
+=================
+
+- Added option to only parsed the listed information of an auction, to skip the rest of the parsing.
+- Fixed wrong type hint in ``ListedAuction`` for ``status``.
+
 3.3.0 (2020-09-09)
 =================